[FIX] Impressao utilizando ids
