+++ jasper_report_backup/models/__init__.py
@@ -4,6 +4,6 @@
 # License LGPL-3 - See http://www.gnu.org/licenses/lgpl-3.0.html
 
 from . import jasper_report
-# from . import jasper_report_parameters
+from . import jasper_report_parameters
 from . import jasper_report_settings
 from . import ir_action_report_xml
