+++ jasper_report_backup/views/jasper_report_parameters_view.xml
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<odoo>
+
+    <!--<record model="ir.ui.view" id="view_jasper_report_parameters_form">-->
+        <!--<field name="name">view.jasper.report.parameters.form</field>-->
+        <!--<field name="model">jasper.report.parameters</field>-->
+        <!--<field name="arch" type="xml">-->
+            <!--<form string="Jasper Report Parameters">-->
+                <!--<sheet>-->
+                    <!--<group>-->
+                        <!--<field name="name"/>-->
+                        <!--<field name="subquery"/>-->
+                    <!--</group>-->
+                <!--</sheet>-->
+            <!--</form>-->
+        <!--</field>-->
+    <!--</record>-->
+
+    <record model="ir.ui.view" id="view_jasper_report_parameters_tree">
+        <field name="name">view.jasper.report.parameters.tree</field>
+        <field name="model">jasper.report.parameters</field>
+        <field name="arch" type="xml">
+            <tree string="Jasper Report Parameters">
+                <field name="name"/>
+                <field name="subquery"/>
+            </tree>
+        </field>
+    </record>
+
+</odoo>
