+++ jasper_report_backup/views/jasper_report_settings.xml
@@ -1,0 +1,65 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<odoo>
+    <!-- DBSource -->
+
+        <record model="ir.ui.view" id="view_jasper_report_db_source_tree">
+            <field name="name">view.jasper.report.db.source.tree</field>
+            <field name="model">jasper.report.db.source</field>
+            <field name="type">tree</field>
+            <field name="arch" type="xml">
+                <tree string="External DB Sources">
+                    <field name="name"/>
+                    <field name="connector"/>
+                </tree>
+            </field>
+        </record>
+
+        <record model="ir.ui.view" id="view_jasper_report_db_source_form">
+            <field name="name">jasper.report.db.source.form</field>
+            <field name="model">jasper.report.db.source</field>
+            <field name="arch" type="xml">
+                <form string="External DB Source">
+                    <sheet>
+                        <group col="1">
+                            <group>
+                                <group>
+                                    <field name="name"/>
+                                    <field name="db_name"/>
+                                    <field name="host"/>
+                                    <field name="port"/>
+                                </group>
+                                <group>
+                                    <field name="connector"/>
+                                    <field name="user_field"/>
+                                    <field name="password" password="True"/>
+                                    <button name="connection_test" string="Test Connection" type="object" icon="fa-refresh"/>
+                                </group>
+                            </group>
+                            <!--<field name="conn_string" invisible="1"/>-->
+                        </group>
+                    </sheet>
+                </form>
+            </field>
+        </record>
+
+        <record model="ir.actions.act_window" id="action_jasper_report_db_source">
+            <field name="name">External Database Sources</field>
+            <field name="res_model">jasper.report.db.source</field>
+            <field name="view_type">form</field>
+            <field name="view_mode">tree,form</field>
+            <field name="view_id" ref="view_jasper_report_db_source_tree"/>
+        </record>
+
+        <menuitem id="jasper_menu" name="Reports" sequence="450"/>
+
+        <menuitem id="jasper_submenu" name="Reports"
+                  parent="jasper_menu" sequence="1"/>
+
+        <menuitem id="jasper_config" name="Settings"
+                  parent="jasper_menu" sequence="2"/>
+
+        <menuitem name="Database Connections" id="menu_db_source"
+                  parent="jasper_config"
+                  action="action_jasper_report_db_source"/>
+
+</odoo>
