+++ jasper_report_backup/views/jasper_report_view.xml
@@ -1,0 +1,85 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<odoo>
+    <record model="ir.ui.view" id="view_jasper_report_tree">
+        <field name="name">view.jasper.report.tree</field>
+        <field name="model">jasper.report</field>
+        <field name="type">tree</field>
+        <field name="arch" type="xml">
+            <tree string="Jasper Reports">
+                <field name="name"/>
+            </tree>
+        </field>
+    </record>
+
+    <record model="ir.ui.view" id="view_jasper_report_form">
+        <field name="name">view.jasper.report.form</field>
+        <field name="model">jasper.report</field>
+        <field name="arch" type="xml">
+            <form string="Jasper Report">
+                <header>
+                    <button string="Build Report" type="object"
+                            name="get_report" class="oe_highlight"/>
+                    <button string="Download Report" type="object"
+                            name="get_file_report"
+                            attrs="{'invisible': [('file_report_binary', '=', False)]}"/>
+
+                </header>
+                <sheet>
+                    <group>
+                        <group>
+                            <field name="name"/>
+                            <field name="db_obj"/>
+                            <field name="template_filename" invisible="1"/>
+                            <field name="template"
+                                   filename="template_filename"/>
+                            <field name="file_report_binary" invisible="1"/>
+                            <field name="output_format"/>
+                        </group>
+
+                    </group>
+                    <notebook>
+                        <page string="Parameters">
+                            <button string="List Parameters"
+                            type="object"
+                            attrs="{'invisible': [('template', '=', False)]}"
+                            name="listing_parameters"
+                            class="oe_highlight"
+                            style="margin: 15px 0"/>
+
+                            <field name="parameters"/>
+
+                        </page>
+                        <page string="Action Reports">
+                            <field name="action_reports" widget="many2many"/>
+                        </page>
+                    </notebook>
+                </sheet>
+            </form>
+        </field>
+    </record>
+
+    <record model="ir.actions.act_window" id="action_jasper_report">
+        <field name="name">Jasper Report</field>
+        <field name="res_model">jasper.report</field>
+        <field name="view_type">form</field>
+        <field name="view_mode">tree,form</field>
+        <field name="view_id" ref="view_jasper_report_tree"/>
+    </record>
+
+    <record model="ir.actions.act_window" id="action_jasper_report_parameters">
+        <field name="name">Jasper Report</field>
+        <field name="res_model">jasper.report.parameters</field>
+        <field name="view_type">form</field>
+        <field name="view_mode">tree,form</field>
+        <!--<field name="view_id" ref="view_jasper_report_tree"/>-->
+    </record>
+
+    <menuitem name="Jasper Report" id="menu_reports"
+              parent="jasper_submenu"
+              action="action_jasper_report"/>
+
+    <menuitem name="Jasper Report Parameters" id="menu_reports_parameters"
+              parent="jasper_submenu"
+              action="action_jasper_report_parameters"/>
+
+</odoo>
